task 0 to 2 modified

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,7 +12,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int  i;
+
 	if (array && size && action)
 	{
 		i = 0;
